+++ lib/features/commit_history/presentation/bloc/commit_history_event.dart
@@ -19,3 +19,7 @@ class SelectCommitFile extends CommitHistoryEvent {
 
   SelectCommitFile({required this.filePath});
 }
+
+class ClearSelectedCommitFile extends CommitHistoryEvent {
+  ClearSelectedCommitFile();
+}