+++ lib/features/files_differences/presentation/bloc/files_differences_bloc.dart
@@ -88,5 +88,14 @@ class FilesDifferencesBloc extends Bloc<FilesDifferencesEvent, FilesDifferencesS
         ),
       );
     });
+
+    on<ClearFileDiff>((event, emit) {
+      emit(
+        state.copyWith(
+          diff: null,
+          selectedFile: null,
+        ),
+      );
+    });
   }
 }