+++ lib/features/repository/presentation/bloc/repository_state.dart
@@ -10,6 +10,8 @@ enum RepositoryBlocStatus {
   cloneSuccess,
   cloning,
   cloneProgress,
+  fetching,
+  fetched,
 }
 
 @MappableClass()