+++ lib/features/repository/presentation/ui/repository_sidebar.dart
@@ -30,7 +30,6 @@ class RepositorySidebar extends StatelessWidget {
         child: Column(
           children: [
             TabBar(
-              isScrollable: true,
               labelPadding: const EdgeInsets.symmetric(horizontal: 12),
               onTap: (index) {
                 final bloc = context.read<RepositoryBloc>();
@@ -47,6 +46,9 @@ class RepositorySidebar extends StatelessWidget {
                       SetRepositoryViewMode(
                         mode: RepositoryViewMode.commitHistory,
                       ),
+                    );
+                    context.read<CommitHistoryBloc>().add(
+                      LoadCommitHistory(),
                     );
                     break;
                   default: