+++ lib/features/working_directory/presentation/bloc/working_directory_event.dart
@@ -64,3 +64,7 @@ class ToggleAllFilesStaging extends WorkingDirectoryEvent {
     required this.stage,
   });
 }
+
+class ClearSelectedFile extends WorkingDirectoryEvent {
+  ClearSelectedFile();
+}