+++ lib/shared/domain/entities/git_commit_entity.dart
@@ -3,11 +3,13 @@ class GitCommitEntity {
   final String author;
   final DateTime date;
   final String message;
+  final bool isUnpushed;
 
   GitCommitEntity({
     required this.sha,
     required this.author,
     required this.date,
     required this.message,
+    required this.isUnpushed,
   });
 }